platform/requirements.txt
fastapi==0.89.0
uvicorn[standard]==0.20.0
firebase-admin==6.0.1
google-cloud-firestore==2.9.1
google-cloud-pubsub==2.14.0